pyproject.toml
[project]
name = "mas-rd-course-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "python-dotenv>=1.2.2",
    "trafilatura>=2.0.0",
    "langchain-google-genai>=4.2.1",
    "langchain-core>=1.2.19",
    "ddgs>=9.11.4",
]
